Hardcode non-Rsc results in update_documentation
It's not like performance numbers obtained for Javac and Scalac are
going anywhere. If we're benchmarking the same software on the same
hardware, and the benchmarking methodology is sufficiently stable
(I believe that ours is), then there's no need to rerun benchmarks
over and over again.

This makes it possible to generate documentation updates via
benchNightly that runs significantly faster than bench.