Make QuickRscTypecheck more or less adequate
10/1 warmup and 10/1 measurement is not enough for HotSpot
to fully warmup - neither on my laptop, nor on our benchbox.

3/10+3/10 seems to be the minimal configuration that reaches
the performance levels of HotRscTypecheck. Unfortunately,
even there run-to-run variance can reach almost a millisecond,
so that is still not completely satisfactory.

As my initial experiments show, changing the number of forks
from one to two significantly reduces run-to-run variance.