Change merge strategy to squash merge
It is too costly to run tests and benchmarks on every commit
of every pull request, so we decided to switch from rebase merge
to squash merge.

I used to dislike squash merge, because I thought that it loses
historical information. However, I was incorrect. The commit produced
by squash merge has a link to the original pull request, so it should
be possible to do full-blown code archaeology if necessary.