Rename unreachable to crash
The name "unreachable" dates back to the early days of Scalameta,
when we introduced it to denote pattern cases that shouldn't happen
but had to be written to satisfy exhaustivity checks.

As we developed Scalameta and, later on, when we started Rsc,
`unreachable` has evolved to more generally denote a control flow path
that shouldn't happen. One of the frequent use cases for `unreachable`
is now asserting that some condition shouldn't evaluate to true,
which derails the original name a little bit and becomes somewhat
confusing.

With the introduction of `unsupported`, confusion has reached its peak.
"Shall I use unreachable? Or maybe unsupported? Is this unreachable
because it's unsupported? Or maybe this is unsupported because it's
unreachable?".

This commit solves this problem once and for all by renaming
`unreachable` to indicate its imperative intent - to crash and burn
the compiler once execution reaches its invocation.