Make bench depend on tests, not the other way around (#69)
I suppose that this was done because tests don't crosscompile to 2.12,
but benchScalac212 does, but that looks really strange.

In this commit, I reversed the direction of dependencies and added
a bit of copy/paste to benchScalac212. Once we fix #62, this copy/paste
won't be necessary anymore, so I'm not too worried about it.