Introduce SemanticDB-based rsc.symtab.Symtab
This commit adds another implementation of a symbol table to Rsc.
This one is based solely on SemanticDB data structures and, as a result,
can interoperate with binary dependencies preprocessed by metacp.

In order to make sure that we're aware of performance implications,
I'm not replacing the old symbol table with the new one right away.
Instead, I'm planning to turn this into a set of small steps, with
every step carefully benchmarked.

This gradual process will require some additional ceremony during
the transition period while both symbol tables coexist. I think this
is well-justified given the benefits of iterative development that
the transition period will enable.