Move stdlibClasspath computation to Build.scala
Computing stdlibClasspath in tests was a bad idea, because dependency
management is something that should be done by build tools.

However, migrating this logic to our build involved a surprising
amount of ceremony. Since metacp is only published for 2.11 and 2.12,
we had to migrate to sbt 1.x, which was fairly laborious. On the flip
side, we can finally invoke metacp directly.