Introduce SemanticDB-based rsc.symtab.Symtab (#70)
* Introduce SemanticDB-based rsc.symtab.Symtab

This commit adds another implementation of a symbol table to Rsc.
This one is based solely on SemanticDB data structures and, as a result,
can interoperate with binary dependencies preprocessed by metacp.

In order to make sure that we're aware of performance implications,
I'm not replacing the old symbol table with the new one right away.
Instead, I'm planning to turn this into a set of small steps, with
every step carefully benchmarked.

This gradual process will require some additional ceremony during
the transition period while both symbol tables coexist. I think this
is well-justified given the benefits of iterative development that
the transition period will enable.

* Move stdlibClasspath computation to Build.scala

Computing stdlibClasspath in tests was a bad idea, because dependency
management is something that should be done by build tools.

However, migrating this logic to our build involved a surprising
amount of ceremony. Since metacp is only published for 2.11 and 2.12,
we had to migrate to sbt 1.x, which was fairly laborious. On the flip
side, we can finally invoke metacp directly.

* Fix eviction warning in plugins.sbt

* Deduplicate V.scalameta

* Temporarily downgrade nativeGC to boehm

Unfortunately, starting with this pull request, immix has started
crashing hot Linux native benchmarks with segmentation faults
(non-hot or non-Linux native benchmarks work just fine).

I am not thrilled about blocking the merge of the new symbol table,
so I'm downgrading the native garbage collector to boehm pending a more
thorough debugging session.

* Address feedback