Laser focus on outlining
Recently, Stu Hood came up with a groundbreaking idea. He suggested
that Rsc can become useful much earlier that we initially thought.
Stu's key insight was to serialize the results of outlining into
a format understood by the Scala compiler and then use that to launch
multiple independent instances of the Scala compiler to parallelize
Scala compilation.

Towards that end, we postponed the initial plan to implement a Scala
typechecker in Rsc and decided to focus on outlining for the time being.
We came up with the following design of a compilation pipeline for Rsc:
  * Metacp converts dependency classpath into the SemanticDB format,
    i.e. the metadata format of the Scalameta toolchain that is used
    internally at Twitter for Scalafix and Language Server Protocol.
  * Rsc computes signatures of all public and protected definitions
    in Scala sources.
  * Rsc saves signatures to disk in SemanticDB format.
  * Mjar converts signatures into the ScalaSignature format,
    i.e. the metadata format of the Scala compiler.

So far, we have implemented support for a subset of Scala to produce
signatures for automatically rewritten core of Twitter Util.
We will be working on adding support for more Scala features according
to our roadmap.

(We have also just learned about an independent effort within the Dotty
project based on a similar architecture by Guillaume Martres.)