Unify classes, traits, interfaces, annotation interfaces and enums
The DefnClass/DefnTrait split was annoying enough, since it barely brought
anything useful to the table, but resulted in duplication (typically, the
logic to handle DefnClass and DefnTrait is mostly if not the same).

Therefore, I wasn't too thrilled about having not just two but five copies
of almost the same thing.