Increase maxColumn from 80 to 100
I have to admit that the maxColumn=80 experiment has failed.
Oftentimes, it is too restrictive and, as a result, every now and then
it forces creation of useless temporary variables.

maxColumn=100 that we're using in Scalameta is much more flexible
without major downsides.

One minor downside is that 100 columns doesn't allow Sublime Text to
horizontally fit a sidebar and two text buffers with the font size
I'm accustomed to. However, I've been working like that in Scalameta,
and it wasn't too much of a problem.