Load Classpath entries directly from filesystem
This slows down benchIndex from 76ms to 101ms on my laptop.

Given that our benchScalasig time is currently 283ms, this slowdown
is just 8% and looks acceptable for the significant simplification
of not having to run Metacp/Metai on our classpaths.

Of course, we don't yet know the cost of dynamically converting
Javac/Scalac metadata to s.SymbolInformation. Neither we know how
much not having to run Metacp will speed up our builds. Therefore,
we cannot yet say whether getting rid of Metacp is a good idea.