Load Classpath infos directly from classfiles
At the moment, Rsc requires that its dependency classpath is expressed
in terms of SemanticDB. This was done this way because:
  1) SemanticDB is easier to work with than ScalaSignatures,
  2) Reading .semanticidx files is faster than scanning the filesystem
     for packages and toplevels.

However, this requirement comes at a steep price. While Rsc itself
emits SemanticDB-compatible classpaths, anything that was compiled by
Scalac or Javac needs to be converted to SemanticDB via Metacp before
being usable on an Rsc dependency classpath. This was fine for our
early experiments, but the bigger the codebases that we're attacking
are, the bigger the hassle of using Metacp.

This commit, together with its parent commit, demonstrate that getting
rid of Metacp is not only feasible but also comes at a reasonable cost.
Overall, if we account for the fact that we no longer need to dump
SemanticDB payloads to disk over the course of normal compilation
(see the child commit for implementation), the performance degradation
for examples/core is only 18%.

                 Before  After
================================
benchIndex       76ms    101ms
benchOutline     138ms   192ms
benchSemanticdb  176ms   231ms
benchScalasig    283ms   349ms