Move parent handling from Outliner to Scheduler
Parent handling is one of the oldest parts of Rsc, and it's been
written during the times when most of the existing technologies
(e.g. synthesizer or sketches) didn't exist yet.

As a result, it's grown increasingly inconsistent with more modern
approaches to synthesis and sketching that were employed for handling
the majority of other language features.

I haven't realized how out of place parent handling is until I noticed
that this is the only user of todo in the entire outliner. Once I
pulled on this thread, that led to a massive refactoring.