Decouple Outliner from Scheduler
After the parent commit moved parent handling from Outliner to Scheduler,
I realized that the only thing in the outliner that still requires a
scheduler is scheduling refinement and existential types.

When I thought about this, I realized that there is little reason for
not scheduling those in advance. Sure, that will require that tpt sketches
are traversed twice - once in scheduler and once in outliner - but the
massive simplification of one of the major compiler modules is worth it.