Replace scopify(Tpt) with scopify(Sketch) (see #395)
This solves a long-standing problem when scopify failures due to
incomplete sketches couldn't be attributed to the blocking sketch.

When such failures occurred, we blocked on Unknown() and hoped that by
the time we run that scopify again, we will have completed the blocking
sketch. This worked for valid programs, but led to some seriously crazy
behavior for some invalid programs.

This pull request implemented the infrastructure for tracking the
sketches, and this commit finally makes use of this infrastructure to
keep track of precise reasons for BlockedResolution.

As a sidenote, in this commit we further observe the tension between
two different usecases for scopify: 1) scopify during outlining when
todo is still in progress, 2) Scopify after outlining when all outlines
have been completed

While there is a significant degree of reuse between these two
algorithms, they are different in that: A) scopify during outlining
can block, B) scopify during outlining can fail. At the moment,
we are reusing the implementation of scopify between usecases 1 and 2,
and as a result the contract for usecase 2 is unnecessarily cumbersome.