Move scopify from Symtab to Outliner
Now that overridesNonNullary is rewritten without scopify, we can finally
move scopify into Outliner, restricting its visibility and applicability
to just the module where it was originally developed.

As explained in 7229e62, I think that scopify shouldn't be reused outside
Outliner. In case we need something similar in semanticdb or scalasig
again, we should consider reimplementing that functionality like it was
done with overridesNonNullary.